requirements.txt
pandas
numpy
torch
matplotlib
scikit-learn
shap
streamlit
plotly
joblib
openpyxl
fastapi
uvicorn
requests
google-generativeai
python-dotenv
